backend/requirements.txt
annotated-types==0.7.0
asttokens==3.0.1
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
decorator==5.2.1
executing==2.2.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
gunicorn==23.0.0
idna==3.11
iniconfig==2.3.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
MIDIUtil==1.2.1
mpmath==1.3.0
networkx==3.6
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
path==17.1.1
pexpect==4.9.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
rich==14.2.0
sentry-sdk==2.46.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
torch==2.9.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.23.0
wcwidth==0.2.14
Werkzeug==3.1.4
